USB: ice40-hcd: Add optimization to improve write throughput
This change clubs the various register read/write in xfer_out in single
SPI message instead of using multiple messages with this we can avoid
unnecessary context switches. This also implements double buffering to make
use of two FIFO's avaialble for efficient bus bandwidth utilizaion of
USB BUS.

CRs-Fixed: 646320

+++ include/trace/events/ice40.h
@@ -282,14 +282,14 @@ TRACE_EVENT(ice40_in,
 
 TRACE_EVENT(ice40_out,
 
-	TP_PROTO(u16 ep, const char *token, u8 len, int ret),
+	TP_PROTO(u16 ep, const char *token, u32 len, int ret),
 
 	TP_ARGS(ep, token, len, ret),
 
 	TP_STRUCT__entry(
 		__field(u16, ep)
 		__string(token, token)
-		__field(u8, len)
+		__field(u32, len)
 		__field(int, ret)
 	),
